fix(util): single top level class warning
