+++ src/main/java/lotto/util/BaseProblem.java
@@ -5,19 +5,3 @@ public interface BaseProblem {
 
     RuntimeException exception();
 }
-
-enum BaseProblemImpl implements BaseProblem {
-    PARSE_LONG_FAILED,
-    PARSE_INTEGER_FAILED,
-    ;
-
-    @Override
-    public String message() {
-        return this.name();
-    }
-
-    @Override
-    public RuntimeException exception() {
-        return Global.EXCEPTION_HANDLER.exception(this);
-    }
-}
