+++ src/main/java/lotto/util/ConsoleProblem.java
@@ -1,0 +1,17 @@
+package lotto.util;
+
+public enum ConsoleProblem implements BaseProblem {
+    PARSE_LONG_FAILED,
+    PARSE_INTEGER_FAILED,
+    ;
+
+    @Override
+    public String message() {
+        return this.name();
+    }
+
+    @Override
+    public RuntimeException exception() {
+        return Global.EXCEPTION_HANDLER.exception(this);
+    }
+}
